backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers==2.2.2
chromadb==0.4.14
PyMuPDF==1.23.7
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.2